+++ contrib/platform/src/com/sun/jna/platform/win32/Kernel32Util.java
@@ -392,23 +392,10 @@ public abstract class Kernel32Util implements WinDef {
                 default:
                     return type;
             }
-        } catch(Win32Exception e) {
-            err = e;
-            throw err;  // re-throw so finally block executed
+        } catch(final Win32Exception e) {
+            throw err = e; // re-throw to avoid return value!
         } finally {
-            try {
-                closeHandle(hFile);
-            } catch(Win32Exception e) {
-                if (err == null) {
-                    err = e;
-                } else {
-                    err.addSuppressedReflected(e);
-                }
-            }
-
-            if (err != null) {
-                throw err;
-            }
+            cleanUp(hFile, err);
         }
     }
 
@@ -944,22 +931,10 @@ public abstract class Kernel32Util implements WinDef {
                 throw new Win32Exception(Kernel32.INSTANCE.GetLastError());
             }
             return QueryFullProcessImageName(hProcess, dwFlags);
-        } catch (Win32Exception e) {
-            we = e;
-            throw we; // re-throw to invoke finally block
+        } catch (final Win32Exception e) {
+            throw we = e; // re-throw to avoid return value!
         } finally {
-            try {
-                closeHandle(hProcess);
-            } catch (Win32Exception e) {
-                if (we == null) {
-                    we = e;
-                } else {
-                    we.addSuppressedReflected(e);
-                }
-            }
-            if (we != null) {
-                throw we;
-            }
+            cleanUp(hProcess, we);
         }
     }
 
@@ -1223,23 +1198,10 @@ public abstract class Kernel32Util implements WinDef {
             }
 
             return modules;
-        } catch (Win32Exception e) {
-            we = e;
-            throw we;   // re-throw so finally block is executed
+        } catch (final Win32Exception e) {
+            throw we = e; // re-throw to avoid return value!
         } finally {
-            try {
-                closeHandle(snapshot);
-            } catch(Win32Exception e) {
-                if (we == null) {
-                    we = e;
-                } else {
-                    we.addSuppressedReflected(e);
-                }
-            }
-
-            if (we != null) {
-                throw we;
-            }
+            cleanUp(snapshot, we);
         }
     }
 
@@ -1293,6 +1255,255 @@ public abstract class Kernel32Util implements WinDef {
             return resultMemory.getWideString(0);
         } else {
             return resultMemory.getString(0);
+        }
+    }
+
+    /**
+     * Gets the priority class of the current process.
+     *
+     * @return The priority class of the current process.
+     * @throws Win32Exception if an error occurs.
+     */
+    public static DWORD getCurrentProcessPriority() {
+        final DWORD dwPriorityClass = Kernel32.INSTANCE.GetPriorityClass(Kernel32.INSTANCE.GetCurrentProcess());
+        if (!isValidPriorityClass(dwPriorityClass)) {
+            throw new Win32Exception(Kernel32.INSTANCE.GetLastError());
+        }
+        return dwPriorityClass;
+    }
+
+    /**
+     * Sets the priority class for the current process.
+     *
+     * @param dwPriorityClass The priority class for the process.
+     * @throws Win32Exception if an error occurs.
+     */
+    public static void setCurrentProcessPriority(final DWORD dwPriorityClass) {
+        if (!isValidPriorityClass(dwPriorityClass)) {
+            throw new IllegalArgumentException("The given priority value is invalid!");
+        }
+        if (!Kernel32.INSTANCE.SetPriorityClass(Kernel32.INSTANCE.GetCurrentProcess(), dwPriorityClass)) {
+            throw new Win32Exception(Kernel32.INSTANCE.GetLastError());
+        }
+    }
+
+    /**
+     * Enables or disables "background" processing mode for the current process
+     *
+     * @param enable If true, enables "background" processing mode, otherwise disables it.
+     * @throws Win32Exception if an error occurs.
+     */
+    public static void setCurrentProcessBackgroundMode(final boolean enable) {
+        // Note: PROCESS_MODE_BACKGROUN_{BEGIN,END} only works with the "current" process handle!
+        final DWORD dwPriorityClass = enable ? Kernel32.PROCESS_MODE_BACKGROUND_BEGIN : Kernel32.PROCESS_MODE_BACKGROUND_END;
+        if (!Kernel32.INSTANCE.SetPriorityClass(Kernel32.INSTANCE.GetCurrentProcess(), dwPriorityClass)) {
+            throw new Win32Exception(Kernel32.INSTANCE.GetLastError());
+        }
+    }
+
+    /**
+     * Gets the priority value of the current thread.
+     *
+     * @return The priority value of the current thread.
+     * @throws Win32Exception if an error occurs.
+     */
+    public static int getCurrentThreadPriority() {
+        final int nPriority = Kernel32.INSTANCE.GetThreadPriority(Kernel32.INSTANCE.GetCurrentThread());
+        if (!isValidThreadPriority(nPriority)) {
+            throw new Win32Exception(Kernel32.INSTANCE.GetLastError());
+        }
+        return nPriority;
+    }
+
+    /**
+     * Sets the priority value for the current thread.
+     *
+     * @param nPriority The priority value for the thread.
+     * @throws Win32Exception if an error occurs.
+     */
+    public static void setCurrentThreadPriority(final int nPriority) {
+        if (!isValidThreadPriority(nPriority)) {
+            throw new IllegalArgumentException("The given priority value is invalid!");
+        }
+        if (!Kernel32.INSTANCE.SetThreadPriority(Kernel32.INSTANCE.GetCurrentThread(), nPriority)) {
+            throw new Win32Exception(Kernel32.INSTANCE.GetLastError());
+        }
+    }
+
+    /**
+     * Enables or disables "background" processing mode for the current thread
+     *
+     * @param enable If true, enables "background" processing mode, otherwise disables it.
+     * @throws Win32Exception if an error occurs.
+     */
+    public static void setCurrentThreadBackgroundMode(final boolean enable) {
+        // Note: THREAD_MODE_BACKGROUND_{BEGIN,END} only works with the "current" thread handle!
+        final int nPriority = enable ? Kernel32.THREAD_MODE_BACKGROUND_BEGIN : Kernel32.THREAD_MODE_BACKGROUND_END;
+        if (!Kernel32.INSTANCE.SetThreadPriority(Kernel32.INSTANCE.GetCurrentThread(), nPriority)) {
+            throw new Win32Exception(Kernel32.INSTANCE.GetLastError());
+        }
+    }
+
+    /**
+     * Gets the priority class of the specified process.
+     *
+     * @param pid Identifier for the running process.
+     * @throws Win32Exception if an error occurs.
+     */
+    public static DWORD getProcessPriority(final int pid) {
+        final HANDLE hProcess = Kernel32.INSTANCE.OpenProcess(WinNT.PROCESS_QUERY_INFORMATION , false, pid);
+        if (hProcess == null) {
+            throw new Win32Exception(Kernel32.INSTANCE.GetLastError());
+        }
+
+        Win32Exception we = null;
+        try {
+            final DWORD dwPriorityClass = Kernel32.INSTANCE.GetPriorityClass(hProcess);
+            if (!isValidPriorityClass(dwPriorityClass)) {
+                throw new Win32Exception(Kernel32.INSTANCE.GetLastError());
+            }
+            return dwPriorityClass;
+        } catch (final Win32Exception e) {
+            throw we = e; // re-throw to avoid return value!
+        } finally {
+            cleanUp(hProcess, we);
+        }
+    }
+
+    /**
+     * Sets the priority class for the specified process.
+     *
+     * @param pid Identifier for the running process.
+     * @param dwPriorityClass The priority class for the process.
+     * @throws Win32Exception if an error occurs.
+     */
+    public static void setProcessPriority(final int pid, final DWORD dwPriorityClass) {
+        if (!isValidPriorityClass(dwPriorityClass)) {
+            throw new IllegalArgumentException("The given priority value is invalid!");
+        }
+
+        final HANDLE hProcess = Kernel32.INSTANCE.OpenProcess(WinNT.PROCESS_SET_INFORMATION, false, pid);
+        if (hProcess == null) {
+            throw new Win32Exception(Kernel32.INSTANCE.GetLastError());
+        }
+
+        Win32Exception we = null;
+        try {
+            if (!Kernel32.INSTANCE.SetPriorityClass(hProcess, dwPriorityClass)) {
+                throw new Win32Exception(Kernel32.INSTANCE.GetLastError());
+            }
+        } catch (final Win32Exception e) {
+            we = e;
+        } finally {
+            cleanUp(hProcess, we);
+        }
+    }
+
+    /**
+     * Gets the priority value of the specified thread.
+     *
+     * @param tid Identifier for the running thread.
+     * @throws Win32Exception if an error occurs.
+     */
+    public static int getThreadPriority(final int tid) {
+        final HANDLE hThread = Kernel32.INSTANCE.OpenThread(WinNT.THREAD_QUERY_INFORMATION, false, tid);
+        if (hThread == null) {
+            throw new Win32Exception(Kernel32.INSTANCE.GetLastError());
+        }
+
+        Win32Exception we = null;
+        try {
+            final int nPriority = Kernel32.INSTANCE.GetThreadPriority(hThread);
+            if (!isValidThreadPriority(nPriority)) {
+                throw new Win32Exception(Kernel32.INSTANCE.GetLastError());
+            }
+            return nPriority;
+        } catch (final Win32Exception e) {
+            throw we = e; // re-throw to avoid return value!
+        } finally {
+            cleanUp(hThread, we);
+        }
+    }
+
+    /**
+     * Sets the priority value for the specified thread.
+     *
+     * @param tid Identifier for the running thread.
+     * @param nPriority The priority value for the thread.
+     * @throws Win32Exception if an error occurs.
+     */
+    public static void setThreadPriority(final int tid, final int nPriority) {
+        if (!isValidThreadPriority(nPriority)) {
+            throw new IllegalArgumentException("The given priority value is invalid!");
+        }
+
+        final HANDLE hThread = Kernel32.INSTANCE.OpenThread(WinNT.THREAD_SET_INFORMATION, false, tid);
+        if (hThread == null) {
+            throw new Win32Exception(Kernel32.INSTANCE.GetLastError());
+        }
+
+        Win32Exception we = null;
+        try {
+            if (!Kernel32.INSTANCE.SetThreadPriority(hThread, nPriority)) {
+                throw new Win32Exception(Kernel32.INSTANCE.GetLastError());
+            }
+        } catch (final Win32Exception e) {
+            we = e;
+        } finally {
+            cleanUp(hThread, we);
+        }
+    }
+
+    /**
+     * Test whether the given priority class is valid.
+     * <p>Intentionally does not accept "background" processing mode flags!</p>
+     *
+     * @param dwPriorityClass The priority value to test.
+     * @return Returns true, if and only if the given priority value was valid
+     */
+    public static boolean isValidPriorityClass(final DWORD dwPriorityClass) {
+        return Kernel32.NORMAL_PRIORITY_CLASS.equals(dwPriorityClass) ||
+            Kernel32.IDLE_PRIORITY_CLASS.equals(dwPriorityClass) ||
+            Kernel32.HIGH_PRIORITY_CLASS.equals(dwPriorityClass) ||
+            Kernel32.REALTIME_PRIORITY_CLASS.equals(dwPriorityClass) ||
+            Kernel32.BELOW_NORMAL_PRIORITY_CLASS.equals(dwPriorityClass) ||
+            Kernel32.ABOVE_NORMAL_PRIORITY_CLASS.equals(dwPriorityClass);
+    }
+
+    /**
+     * Test whether the given thread priority is valid.
+     * <p>Intentionally does not accept "background" processing mode flags!</p>
+     *
+     * @param nPriority The priority value to test.
+     * @return Returns true, if and only if the given priority value was valid
+     */
+    public static boolean isValidThreadPriority(final int nPriority) {
+        switch(nPriority) {
+            case Kernel32.THREAD_PRIORITY_IDLE:
+            case Kernel32.THREAD_PRIORITY_LOWEST:
+            case Kernel32.THREAD_PRIORITY_BELOW_NORMAL:
+            case Kernel32.THREAD_PRIORITY_NORMAL:
+            case Kernel32.THREAD_PRIORITY_ABOVE_NORMAL:
+            case Kernel32.THREAD_PRIORITY_HIGHEST:
+            case Kernel32.THREAD_PRIORITY_TIME_CRITICAL:
+                return true;
+            default:
+                return false;
+        }
+    }
+
+    private static void cleanUp(final HANDLE h, Win32Exception we) {
+        try {
+            closeHandle(h);
+        } catch (final Win32Exception e) {
+            if (we == null) {
+                we = e;
+            } else {
+                we.addSuppressedReflected(e);
+            }
+        }
+        if (we != null) {
+            throw we;
         }
     }
 }